Refactor Subsystem Code
