+++ src/main/java/frc/team4909/robot/subsystems/DriveTrainSubsystem.java
@@ -1,4 +1,4 @@
-package frc.team4909.robot.subsystems.drivetrain;
+package frc.team4909.robot.subsystems;
 
 import com.revrobotics.CANSparkMax;
 import com.revrobotics.CANSparkMaxLowLevel.MotorType;
